+++ crates/war/src/darksiders1/code/vigil/gfc/mod.rs
@@ -6,3 +6,4 @@ pub mod reflection;
 pub mod save;
 pub mod script;
 pub mod util;
+pub mod world;