+++ crates/war/src/darksiders1/code/vigil/gfc/util/mod.rs
@@ -1,3 +1,4 @@
 pub mod crc;
+pub mod directorymanager;
 pub mod hstring;
 pub mod hstringmanager;