+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.0 – 2019-10-28
+
+### Added
+
+- A command to unpack the resource packages.
+- A command to display the hash of a string.
+
 ## v0.1.1 – 2019-09-27
 
 ### Fixed